requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
google-generativeai==0.8.4
google-ai-generativelanguage==0.6.15
grpcio-status==1.63.0rc1
protobuf==4.25.6